+++ components/automatic-grid/demo/demo.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <link rel="stylesheet" href="../../grid/dist/grid.production.min.css">
-<link rel="stylesheet" href="../components-theme.css">
+<link rel="stylesheet" href="../coherent-gameface-components-theme.css.css">
 <link rel="stylesheet" href="../style.css">
 
 <style>
