+++ components/checkbox/demo/demo.html
@@ -5,7 +5,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../components-theme.css">
+    <link rel="stylesheet" href="../coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="../style.css">
     <title>Document</title>
 </head>
