+++ components/dropdown/README.md
@@ -117,7 +117,7 @@ These are some of the differences to the standard select that are not yet suppor
 ## Add the Styles
 
 ~~~~{.css}
-<link rel="stylesheet" href="components-theme.css">
+<link rel="stylesheet" href="coherent-gameface-components-theme.css">
 <link rel="stylesheet" href="style.css">
 <link rel="stylesheet" href="node_modules/coherent-gameface-scrollable-container/style.css">
 <link rel="stylesheet" href="node_modules/coherent-gameface-slider/styles/horizontal.css">
