+++ components/dropdown/script.js
@@ -6,7 +6,7 @@
 import components from 'coherent-gameface-components';
 import ScrollableContainer from 'coherent-gameface-scrollable-container';
 import template from './template.html';
-import theme from '../../theme/components-theme.css';
+import theme from '../../theme/coherent-gameface-components-theme.css';
 import style from './style.css';
 
 const KEYCODES = components.KEYCODES;
