+++ components/menu/demo/demo.html
@@ -3,7 +3,7 @@
 <html lang="en">
 
 <head>
-    <link rel="stylesheet" href="../components-theme.css">
+    <link rel="stylesheet" href="../coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="../menus/menu.css">
     <link rel="stylesheet" href="../menus/bottom/bottom.css">
     <link rel="stylesheet" href="../menus/left/left.css">
