+++ components/modal/demo/demo.html
@@ -3,7 +3,7 @@
 <html lang="en">
 
 <head>
-    <link rel="stylesheet" href="../components-theme.css">
+    <link rel="stylesheet" href="../coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="../style.css">
     <style>
         .character-customization {
