+++ components/progress-bar/coherent-gameface-components-theme.css
