+++ components/radial-menu/README.md
@@ -138,7 +138,7 @@ opened again.
 ## Add the Styles
 
 ~~~~{.css}
-<link rel="stylesheet" href="components-theme.css">
+<link rel="stylesheet" href="coherent-gameface-components-theme.css">
 <link rel="stylesheet" href="style.css">
 ~~~~
