+++ components/router/demo/demo.html
@@ -1,7 +1,7 @@
 <!--Copyright (c) Coherent Labs AD. All rights reserved. -->
 <!DOCTYPE html>
 <html lang="en">
-    <link rel="stylesheet" href="../components-theme.css">
+    <link rel="stylesheet" href="../coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="../style.css">
     <style>
         .menu {
