+++ components/scrollable-container/demo/demo.html
@@ -2,7 +2,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <link rel="stylesheet" href="../components-theme.css">
+    <link rel="stylesheet" href="../coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="../node_modules/coherent-gameface-slider/styles/horizontal.css">
     <link rel="stylesheet" href="../node_modules/coherent-gameface-slider/styles/vertical.css">
     <link rel="stylesheet" href="../style.css">
