+++ components/slider/coherent-gameface-components-theme.css
