+++ components/slider/demo/demo.html
@@ -2,7 +2,7 @@
 <!DOCTYPE html>
 <html lang="en">
     <head>
-        <link rel="stylesheet" href="../components-theme.css">
+        <link rel="stylesheet" href="../coherent-gameface-components-theme.css">
         <link rel="stylesheet" href="../styles/horizontal.css">
         <link rel="stylesheet" href="../styles/vertical.css">
         <style>
