+++ components/tabs/README.md
@@ -105,7 +105,7 @@ document.querySelector('[data-name="panel"]').appendChild(panel);
 ## Add the Styles
 
 ~~~~{.css}
-<link rel="stylesheet" href="components-theme.css">
+<link rel="stylesheet" href="coherent-gameface-components-theme.css">
 <link rel="stylesheet" href="styles.css">
 ~~~~
