+++ docs/content/examples/checkbox.html
@@ -11,7 +11,7 @@ draft: false
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/components/components-theme.css">
+    <link rel="stylesheet" href="/components/coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="/components/checkbox/style.css">
     <title>Document</title>
     <style>
