+++ docs/content/examples/menu.html
@@ -9,7 +9,7 @@ draft: false
 <html lang="en">
 
 <head>
-    <link rel="stylesheet" href="/components/components-theme.css">
+    <link rel="stylesheet" href="/components/coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="/components/menu/menu.css">
     <link rel="stylesheet" href="/components/menu/bottom.css">
     <link rel="stylesheet" href="/components/menu/left.css">
