+++ docs/content/examples/radial-menu.html
@@ -9,7 +9,7 @@ draft: false
 <html lang="en">
 
 <head>
-	<link rel="stylesheet" href="/components/components-theme.css">
+	<link rel="stylesheet" href="/components/coherent-gameface-components-theme.css">
 	<link rel="stylesheet" href="/components/radial-menu/style.css">
 	<style>
 		body {
