+++ docs/content/examples/router.html
@@ -7,7 +7,7 @@ draft: false
 <!--Copyright (c) Coherent Labs AD. All rights reserved. -->
 <!DOCTYPE html>
 <html lang="en">
-<link rel="stylesheet" href="/components/components-theme.css">
+<link rel="stylesheet" href="/components/coherent-gameface-components-theme.css">
 <link rel="stylesheet" href="/components/router/style.css">
 <style>
     .page-wrapper {
