+++ docs/content/examples/slider.html
@@ -9,7 +9,7 @@ draft: false
 <html lang="en">
 
 <head>
-    <link rel="stylesheet" href="/components/components-theme.css">
+    <link rel="stylesheet" href="/components/coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="/components/slider/horizontal.css">
     <link rel="stylesheet" href="/components/slider/vertical.css">
     <style>
