+++ docs/content/examples/tabs.html
@@ -12,7 +12,7 @@ draft: false
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <link rel="stylesheet" href="/components/components-theme.css">
+    <link rel="stylesheet" href="/components/coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="/components/tabs/style.css">
     <style>
         .page-wrapper {
