+++ index.html
@@ -4,9 +4,8 @@
 
 <head>
     <link rel="stylesheet" href="demo.css">
-    <link rel="stylesheet" href="theme/components-theme.css">
+    <link rel="stylesheet" href="theme/coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="components/grid/dist/grid.production.min.css">
-    <link rel="stylesheet" href="components/menu/components-theme.css">
     <link rel="stylesheet" href="components/menu/menus/menu.css">
     <link rel="stylesheet" href="components/menu/menus/bottom/bottom.css">
     <link rel="stylesheet" href="components/menu/menus/left/left.css">
