+++ samples/user_interface/main/index.html
@@ -6,7 +6,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <link rel="stylesheet" href="node_modules/coherent-gameface-menu/components-theme.css">
+    <link rel="stylesheet" href="node_modules/coherent-gameface-menu/coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="node_modules/coherent-gameface-grid/dist/grid.production.min.css">
     <link rel="stylesheet" href="node_modules/coherent-gameface-menu/menus/menu.css">
     <link rel="stylesheet" href="node_modules/coherent-gameface-menu/menus/bottom/bottom.css">
