+++ samples/user_interface/settings/settings.html
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
     <link rel="stylesheet" href="node_modules/coherent-gameface-grid/dist/grid.production.min.css">
-    <link rel="stylesheet" href="node_modules/coherent-gameface-checkbox/components-theme.css">
+    <link rel="stylesheet" href="node_modules/coherent-gameface-checkbox/coherent-gameface-components-theme.css">
     <link rel="stylesheet" href="node_modules/coherent-gameface-checkbox/style.css">
     <link rel="stylesheet" href="node_modules/coherent-gameface-modal/style.css">
     <link rel="stylesheet" href="node_modules/coherent-gameface-tabs/style.css">
