+++ samples/user_interface/shop/shop.html
@@ -6,7 +6,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Shop</title>
-	<link rel="stylesheet" href="node_modules/coherent-gameface-modal/components-theme.css">
+	<link rel="stylesheet" href="node_modules/coherent-gameface-modal/coherent-gameface-components-theme.css">
 	<link rel="stylesheet" href="node_modules/coherent-gameface-modal/style.css">
     <link rel="stylesheet" href="node_modules/coherent-gameface-grid/dist/grid.production.min.css">
 	<script src="node_modules/coherent-gameface-components/umd/components.development.js"></script>
